Add a library
Add a library to select which book you want to view the graph for and also fix the links in the navigation bar.

Fixes: #95.
Part-of: #90.
