+++ .github/workflows/nlp-timeline-gen-workflow.yml
@@ -12,7 +12,7 @@ jobs:
         with:
           python-version: '3.10'
       - name: Install backend dependencies
-        run: ./scripts/install_backend_dependencies.sh
+        run: ./scripts/setup_backend.sh
       - name: Lint python files with flake8
         run: ./scripts/check_py_format.sh
       - name: Lint web related files with prettier
@@ -28,7 +28,7 @@ jobs:
       - uses: actions/setup-node@v3
       - name: Install frontend dependencies
         run: |
-          ./scripts/install_frontend_dependencies.sh
+          ./scripts/setup_frontend.sh
       - name: Build
         run: |
           cd client
