+++ .gitignore
@@ -9,6 +9,7 @@ __pycache__/
 # Distribution / packaging
 .Python
 build/
+/client/public/library
 develop-eggs/
 dist/
 downloads/
