+++ client/package.json
@@ -13,7 +13,8 @@
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.2.0",
     "react-force-graph": "^1.41.14",
-    "react-scripts": "^2.1.3"
+    "react-scripts": "^2.1.3",
+    "react-router-dom": "^6.4.3"
   },
   "scripts": {
     "predeploy": "npm run build",
