+++ client/public/ChronoLogoMini.png
