+++ client/src/About.js
@@ -1,0 +1,13 @@
+import React, { useEffect } from "react";
+import "./App.css";
+import "./Library.css";
+
+function About() {
+  useEffect(() => {
+    document.body.style.backgroundColor = "#eae0d5";
+  });
+
+  return <div className="about"></div>;
+}
+
+export default About;
