+++ client/src/ChronoLogNavBar.js
@@ -9,34 +9,24 @@ function ChronoLogNavBar() {
   return (
     <Navbar className="color-nav" variant="dark" expand="lg">
       <Container>
-        <Navbar.Brand className="nav-brand" href="#home">
+        <Navbar.Brand className="nav-brand" href="/">
           ChronoLog.
         </Navbar.Brand>
         <Navbar.Toggle aria-controls="basic-navbar-nav" />
         <Navbar.Collapse id="basic-navbar-nav">
           <Nav className="me-auto">
-            <Nav.Link className="nav-link" href="#home">
+            <Nav.Link className="nav-link" href="/about">
               About
             </Nav.Link>
-            <NavDropdown title="Library" id="basic-nav-dropdown">
-              <NavDropdown.Item href="#action/2.1">
-                Winnie The Pooh
-              </NavDropdown.Item>
-              <NavDropdown.Item href="#action/2.2">
-                Little Women
-              </NavDropdown.Item>
-              <NavDropdown.Item href="#action/2.3">
-                Harry Potter and the Philosopher's Stone
-              </NavDropdown.Item>
-            </NavDropdown>
+            <Nav.Link className="nav-link" href="/library">
+              Library
+            </Nav.Link>
             <NavDropdown title="Developers" id="basic-nav-dropdown">
-              <NavDropdown.Item href="#action/3.1">GitHub</NavDropdown.Item>
-              <NavDropdown.Item href="#action/3.2">
-                Documentation
+              <NavDropdown.Item href="https://github.com/JamesLee-Jones/ChronoLog">
+                GitHub
               </NavDropdown.Item>
-              <NavDropdown.Divider />
-              <NavDropdown.Item href="#action/3.4">
-                Separated link
+              <NavDropdown.Item href="https://github.com/JamesLee-Jones/ChronoLog/blob/main/README.md">
+                Documentation
               </NavDropdown.Item>
             </NavDropdown>
           </Nav>
