+++ scripts/setup.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
-./scripts/install_frontend_dependencies.sh
-./scripts/install_backend_dependencies.sh
+./scripts/setup_frontend.sh
+./scripts/setup_backend.sh
