+++ scripts/setup_frontend.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
 
+cp -R timelines client/public/library
 cd client
 npm install
